requirements.txt
Pillow
py-zabbix
pyyaml.yandex

